feat: add auth banner to profile & squad pages
- extract out the banner component into the shared lib
